refactor: split frame and mux modules
Separate frame and mux concepts into focused modules so lib.rs files act as public re-export surfaces instead of holding every implementation detail.

This improves locality for frame sources, Domain adapters, Pane Views, client refresh, and history while preserving the public crate interface.
